conf: Allow > UINT_MAX of cache for NUMA nodes
The high-bandwidth memory (HBM) in cache mode might be greater than
UINT_MAX of cache per NUMA node, so change to unsigned long long.

+++ src/conf/numa_conf.h
@@ -263,7 +263,7 @@ void virNumaDistanceFormat(virBuffer *buf,
 typedef struct _virNumaCache virNumaCache;
 struct _virNumaCache {
     unsigned int level; /* cache level */
-    unsigned int size;  /* cache size */
+    unsigned long long size;  /* cache size */
     unsigned int line;  /* line size, !!! in bytes !!! */
     virNumaCacheAssociativity associativity; /* cache associativity */
     virNumaCachePolicy policy; /* cache policy */
